[Substrait] Support root case in PlanRel message. (iree-org/iree-llvm-sandbox#821)
This essentially consist in adding optional field names to the
`PlanRelOp`. Since verifying if those field names match the yielded type
of the relation consists of the same logic as the one used in the
verifier of the `NamedTableOp`, this commit also factors out that
verification log. While touching that code, the commit also slightly
extends the error messages emitted on verification failure.

Signed-off-by: Ingo Müller <ingomueller@google.com>