[Substrait] Implement first version of FilterOp and Expression ops. (iree-org/iree-llvm-sandbox#812)
* [Substrait] Implement first version of FilterOp and Expression ops.

This corresponds to the `FilterRel` message and the ones contained in
it. Since one of the messages is an `Expression` message, this PR also
introduces the `ExpressionOpInterface` as well as the `LiteralOp` as
well as the basic import/export infrastructure for such ops/messages.
The `LiteralOp` corresponds to the `Expression.Literal` message type,
and the `ExpressionOpInterface` encapsualtes common properties and
functionality of all types of `Expression` messages. Finally, the PR
also adds attribute constraints for typed attributes of the currently
supported types.

Signed-off-by: Ingo Müller <ingomueller@google.com>