[Substrait] Extend emit deduplication to FilterOp.  (iree-org/iree-llvm-sandbox#832)
This adds a new pattern that makes emit deduplication to `filter` ops by
adding a new pattern. This pattern is similar to the previous one in
that it adds a "deduplicating" `emit` op in front of the root `filter`
op and one that re-establishes the original order after it but has the
additional complexity of having to update the region to work on the
deduplicated element type as well.

Signed-off-by: Ingo Müller <ingomueller@google.com>