[Substrait] Simplify assembly format of `field_reference` op. (iree-org/iree-llvm-sandbox#849)
This PR removes a pair of `[.]` around the integer array `position`
attribute of the `field_reference` op, which isn't necessary. The
original format had used the additional pair of brackets out of
ignorance and was misleading as it suggested that there is an additional
nesting level, which actually doesn't exist.

Signed-off-by: Ingo Müller <ingomueller@google.com>