Rename everything related to "structured" to "Substrait (MLIR)".
This PR renames everything (variable names, file names, target names,
comments, etc.) related to "structured" to either "Substrait" or
"Substrait MLIR". In order to avoid confusion or name clashes with
Substrait proper, usually "Substrait MLIR" was used, such as in the
Python package name `substrait_mlir`. However, in some contexts, the
relationship to MLIR is clear, so the MLIR suffix was omitted, such as
in the executable name `substrait-opt`. Other than that, the changes are
mostly mechanic, with a small number of exceptions where comments had to
be rephrased somewhat.

Signed-off-by: Ingo Müller <ingomueller@google.com>