Move (internal) path of `llvm-project` git submodule.
This is determined by a property set in `.gitmodules` and only affects
where git stores the metadata of that submodule, not the checked out
files. This doesn't affect the functionality of the code and probably
nobody would ever notice but there is also no reason to keep this
historic path.

Signed-off-by: Ingo Müller <ingomueller@google.com>