fix: use JSON notatition for SPDX identifier
This is necessary because the action expects the license identifier to
be a space-separated list of strings via its main parameter mechanism,
which breaks with our license identifier, which contains strings.
Fortunately, there is a fallback mechanism via environment variables
that allows to specify the list of identifiers as a JSON array.

Signed-off-by: Ingo Müller <ingomueller@google.com>