feat: implement `AdvancedExtension` and at it to `Plan`
This PR implements the `AdvanceExtension` message type as a custom
attribute with two `StringAttr`s as well as an `AnyType` with a
parameter for the type URL to represent the `Any` messages as typed
`StringAttr`s. The PR also adds those attributes to the `PlanOp` in
order to represent the `advanced_extensions` message of the `Plan`
message.

We do not currently need this attribute but it is one of the few missing
fields of the `Plan` message as well as many other messages we have
otherwise implemented, so it allows us to get to higher coverage with
relatively little effort.

Signed-off-by: Ingo Müller <ingomueller@google.com>