chore: move attribute definitions to dedicated file
This PR moves all attribute definitions from `SubstraitTypes.td` to a
new file `SubstraitAttrs.td` in order to separate the definitions of the
two things better. Having them in one file was acceptable for boot
strapping but the file has grow significantly lately, so I think it's
time to clean up a bit.

Signed-off-by: Ingo Müller <ingomueller@google.com>