feat: implement `extension_table` op
This PR implements the `extension_table` op, the MLIR equivalent of the
`ReadRel.ExtensionTable` message. Since that message uses the
`google.protobuf.Any` message type, the PR also slightly extends and
improves the handling of that message in the dialect. Finally, because
`extension_table` is the second op corresponding to a `ReadRel` case
(after `named_table`), the PR makes some effort to factor out common
logic between the two, namely, how the named structs representing the
schema of the op are handled. However, there is still some opportunity
for factoring out further that the PR does not do, such as defining a
`ReadRelInterface`.

Signed-off-by: Ingo Müller <ingomueller@google.com>