fix: handle missing expression type in protobuf import correctly (#71)
This PR fixes the handling of `Expression` protobuf messages that do not
of the `expression` type set. Such a message would be invalid Substrait
but valid protobuf, so the import logic has to be able to handle that
case gracefully. The PR now has a dedicated `case` for that case.

Signed-off-by: Ingo Müller <ingomueller@google.com>