feat: elide attribute type with new `StaticallyTypedAttrInterface` (#59)
This allows to omit the type of an attribute from the assembly if it can
be inferred from the attribute value. For example, the type is redundant
in `#substrait.timestamp<100us> : !substrait.timestamp`. The built-in
`TypedAttrInterface`, however, forces the appearance of the type in the
assembly. The new interface is almost identical but does not enforce it.
The PR also makes the two timestamp, the date, and the time attributes
implement that interface.

Signed-off-by: Ingo Müller <ingomueller@google.com>