chore: fix `FloatType` building and refactor related export logic
The way `FloatType`s are built was changed upstream: instead of
`FloatType::getF32`, we now use `Float32Type::get`. This commit adapts
the usages of the previous pattern. While touching that code, the commit
also refactors both `exportFloatType` and `exportIntegerType` to use a
`switch` statement on the type width instead of constructing all
possible MLIR types and comparing linearly against them, which is less
efficient.

Signed-off-by: Ingo Müller <ingomueller@google.com>