feat: implement `aggregation_phase` of `call` op
This PR adds the `aggregation_phase` to the `call` op, which corresponds
to the `phase` field of the `AggregationFunction` message. The PR also
revisits the assembly format of the `aggregation_invocation` attribute
of the `call` op and now handles the two jointly such that either of the
two can be ommitted if it has the default value. Finally, the PR
revisits the LIT tests such that both aggregation detauls are tested
jointly and explicitly instead of being piggy-backed onto the remaining
tests. This is the last non-type feature we need to run TPC-H Q6 as
produced by Isthmus.