feat: implement `Plan.expected_type_urls` field (#67)
This PR adds support for the `expected_type_urls` field of the `Plan`
message. This is pretty straight-forward: the repeated field of string
becomes an `ArrayAttr` of `StringAttr`s. We don't really need this field
currently but it's the last missing field of the `Plan` message, so it
allows us to tick that message type off.

Signed-off-by: Ingo Müller <ingomueller@google.com>