feat: add `aggregate` op and extend functions accordingly (#70)
* feat: add `aggregate` op and extend functions accordingly

This PR adds support for the `AggregateRel` from the Substrait spec in
the form of the `aggregate` op. This is arguably the most complex op
implemented so far. It has an optional enum argument that requires
custom parsing, several optional regions that require custom parsing, an
attribute that depends on the presence and contents of the regions and
requires custom parsing to omit it in the common case, and return types
that depend on the two regions and the attribute. What's more, the
current version of the spec is such that it is almost impossibly to
interpret "grouping sets" because it relies on protobuf message
equality, which is something can protobuf does not offer. The current
implementation, thus, implements a best effort by using op equality
instead (but needs to run CSE during export to ensure op uniqueness).
The PR also extends the `call` op to represent also `AggregateFunction`
messages (in addition to `ScalarFunction` messages), which are used by
the new `aggregate` op. Finally, the PR replaces some usages of
`UnknownLoc` with the location from the `ImplicitOpLocBuilder`
wherever one is available (including places there aren't otherwise
affected by this PR).

Signed-off-by: Ingo Müller <ingomueller@google.com>