feat: implement support for the `PlanVersion` message type (#69)
This PR introduces support for the `PlanVersion` message type. This is
done by introduces the `VersionAttr`, which corresponds to the `Version`
message type, which, in turn, is the type of the only field of
`PlanVersion`.

Since `PlanVersion` is a top-level message, this PR extends the import
mechanism: when reading a protobuf message, we need to know which type
it is, so when we translate a protobuf message into MLIR, we need to
specify which message type the input is. The PR, thus, adds a new flag
to the `substrait-translate` tool, together with a corresponding
top-level translation function. This also slightly changed some existing
error messages and the corresponding tests.

This PR does not yet factor out the handling of the version of the
`Plan` message type. That should follow up soon but, since the new
`VersionAttr` has a slightly different assembly format than the
corresponding part of the `PlanOp`, that change will require to modify
all test cases, which would bloat this PR.

Signed-off-by: Ingo Müller <ingomueller@google.com>