feat: update LLVM to latest version as of Jan. 31, 2025 (956cfa69b15) (#74)
* feat: update LLVM to latest version as of Jan. 31, 2025 (956cfa69b15)

This PR updates the LLVM git submodule to the latest version as of
January 31, 2025, namely llvm/llvm-project@956cfa69b15 and adapts
the code base to the changes since the last update:

* remove unused and remove dependency to `mlir-cpu-runner`
* chore: adapt to renamed `applyPatternsGreedily`
* chore: fix `FloatType` building and refactor related export logic

The way `FloatType`s are built was changed upstream: instead of
`FloatType::getF32`, we now use `Float32Type::get`. This commit adapts
the usages of the previous pattern. While touching that code, the commit
also refactors both `exportFloatType` and `exportIntegerType` to use a
`switch` statement on the type width instead of constructing all
possible MLIR types and comparing linearly against them, which is less
efficient.

Signed-off-by: Ingo Müller <ingomueller@google.com>