feat: introduce `interval_year` and `interval_day` types (#48)
This PR introduces two new atomic interval types:

Interval Year to Month: Supports a range of [-10,000..10,000] years, with
month-level precision. Typically stored as separate integers for years and
months (e.g., 12y 3m), but only the total number of months is considered
significant. For example, 1y 0m is equivalent to 0y 12m or 1001y -12000m.

Interval Day to Second: Supports a range of [-3,650,000..3,650,000] days, with
second precision. Typically stored as separate integers for days and seconds,
but only the total number of seconds is considered significant. For example,
1d 0s is equivalent to 0d 86400s.