chore: update LLVM submodule to latest as of 02/19/2025 (0de2ccab7b) (#88)
This PR updates the LLVM submodule to llvm/llvm-project@0de2ccab7b, the
latest version as of today. That version contains a fix for `nanobind`'s
`stubgen` (in llvm/llvm-project#127518), which we require for the typing
stubs for #78.

This PR requires to update the git modules in existing git clones:

```bash
git submodule update --recursive
```

Signed-off-by: Ingo Müller <ingomueller@google.com>