chore: update substrait-cpp submodule to latest version (904a493fc) (#92)
This PR updates the git submodule of substrait/substrait-cpp to the currently
latest version, substrait/substrait-cpp@904a493fc. The previsouly tagged
version of substrait-cpp was ~4 months old. The latest version includes
several fixes to be able to build on Windows - which in turn, is a prerequisite
for getting `substrait-mlir` to build on Windows.