build: use and adapt to more recent Protobuf from `substrait-cpp` (#99)
This PR upgrades the codebase to the more recent Protobuf version
downloaded and compiled by `substrait-cpp`. Previously, we used
`find_package` from our CMake files, even though we don't use those
targets directly and only use those imported by `substrait-cpp` already.
Also, since recently, `substrait-cpp` can download Protobuf if none is
installed in the systems, so we rely on that mechanism in CI now. Since
that version is newer than the one we used in CI until now, the PR also
updates the code base to the new version.

The PR suppresses some warnings in the headers of the new Protobuf
version, which are errors in CI due to our using of `-Werror`, by
setting some diagnostic pragmas. However, it would be better if those
headers were imported as `SYSTEM` headers by `substrait-cpp`, but I
still need to figure out how to do that and update that repository
accordingly.

Signed-off-by: Ingo Müller <ingomueller@google.com>