test: add tooling for new `Round-Trip` test type (#109)
This PR adds some infrastructure for testing that Substrait plans
round-trip through `substrait-mlir` without loosing any information.
This is currently based on the JSON format and consist in comparing (1)
a normalized version of the input plan with (2) a normalized version of
the plan after round-tripping it through `substrait-mlir`. JSON is maybe
the format that makes this easiest and most robust since normalizing
JSON can be done with mature tools (deserialize from and reserialize to
JSON). The PR also extends LIT to run these tests and adds a test case
that exercises the round-trip testing infrastructure on a couple of
plans based on the `plan.mlir` cases from the current dialect tests.

#111 outlines a possibly superior approach that we can use in the future.

Signed-off-by: Ingo Müller <ingomueller@google.com>