build: remove wrong and unused `ibis` Python requirement
Turns out that that package has nothing to do with Apache Ibis, which we
use in some integration tests. It's some templating engine that happens
to have the same name and must have ended up in `requirements.txt` by
accident. Since recent CI runs broke due to not finding that package and
we don't need it, this PR removes that package.

Signed-off-by: Ingo Müller <ingomueller@google.com>