feat: implement script that produces stats on implementation coverage (#56)
This PR adds a script that computes statistics on how much
of the Substrait specification is currently covered by this project
based on which of the protobuf fields of the spec occur in the
"import" test cases. The PR also adds a new CI target that
checks whether the markdown file holding the spec is
up-to-date and fails with an error message indicating how to
update that file in case that's necessary.

Signed-off-by: Ingo Müller <ingomueller@google.com>