fix: print make command instead of executing it in failure message (#130)
This PR fixes the GH workflow that ensures that the coverage stats are
up-to-date. The workflow is supposed to print a message in case the
stats need to be updated including the command that the developer should
run to update them. Due to wrong quoting, that command was actually
executed rather than being printed, so this PR changes to quotes to
protect the command.

Signed-off-by: Ingo Müller <ingomueller@google.com>