build: add direct CMake dependencies and order alphabetically
While preparing the BUILD files for bazel, I ran into missing direct
dependencies. Bazel can notice if a build target uses a particular
dependency but only declares that dependency *indirectyl*, i.e., there
is a dependency a -> b -> c, even though a uses c directly. This often
works but may break if the dependencies of the intermediate target
change, so it is better to always declare all dependencies directly.

Signed-off-by: Ingo Müller <ingomueller@google.com>