style: enable `clang-tidy`'s `misc-include-cleaner` check
This may seem in contrast with the LLVM Coding Standards, which mandates
to "include as little as possible." However, it does say that "[we]
**must** include all of the header files that [we] are using," though it
gives the freedom to do that "either directly or indirectly through
another header file." The rule is, in fact, mainly targeted at forward
declarations of classes that aren't "used" in the C++ sense with the
goal of reducing compile time. Using the `misc-include-cleaner`, thus,
seems like a good compromise: it allows to enforce removing unused
headers, which improves compile time, at the expense of not being able
to indirectly include other headers anymore, which is arguably a minor
disadvantage.

This PR also adds a CI check that enforces the correct include style
(`"." vs `<.>`) for the external libraries that we are using.