style: reconfigure `clang-format`'s `IncludeBlocks` to `Regroup` (#144)
Until now, we have basically used "grouped" includes manually but
`clang-format` did not enforce the underlying categories, so some files
did not completely adhere to that style. The new style also defines the
`substrait-mlir` headers as its own category and, following the spirit
of the LLVM Coding Standards to have more specific headers first, keeps
those in a dedicated block after the local headers and before the
LLVM/MLIR headers.

Signed-off-by: Ingo Müller <ingomueller@google.com>