build: enable `readability-identifier-naming` check of `clang-tidy` (#141)
That flag was disabled when `clang-tidy` was set up initially in #140
because it failed in too many places. This PR enables it and applies the
automatic fixes such that identifier naming will be enforced in the
future.

Two variables were called `struct_`, which `clang-tidy` wanted to rename
to `struct`, which is a keyword so the renaming fails. I, thus, manually
renamed these variables to `structMsg` (because the hold a protobuf
message).

Signed-off-by: Ingo Müller <ingomueller@google.com>