test: unbuffer Python with CLI flag rather than lit config (#139)
This PR moves the disabling of Python's output buffering to a `-u` flag
in the `RUN` line of each individual lit test. Without this
configuration, the output on stdout from Python gets delayed with
respect to the output from native libraries (such as the MLIR Python
bindings) such that the order is not preserved and `FileCheck` checks
fail. Previously, the configuration was don in `lit.cfg.py` but that
file may not be used by other test runners.

Signed-off-by: Ingo Müller <ingomueller@google.com>