feat: implement real `relation` type (#78)
This PR introduces a new RelationType as the operands and result types
of RelOps, where previously tuple had been used as a placeholder. While
this changes little in the structure of the IR, it (1) underlines the
subtle difference between the relation, a container of records, and the
records themselves, which had both been represented as tuples before
and (2) it allows defining an even shorter type name, rel, for that
type. The latter is implemented using a custom printer and parser for
these "pseudo type names" (which are really keywords for the printer
and parser), which lays the basis for similar short-hand forms for some
of our other types with long type names.

Signed-off-by: Ingo Müller <ingomueller@google.com>