feat: implement short-hand version for remaining types
This commits adds a short-hand version for the remaining types in the
Substrait dialect and adapts the tests accordingly. This means that we
can now write `interval_ds` instead of `!substrait.interval_day_second`,
for example. The short-hand versions mainly consist of just the op name
(e.g., `date` instead of `!substrait.date` but a few short-hand versions
are shortend further (like `interval_ds`). The short-hand versions can
be used everywhere were the custom `SubstraitType` printer and parser is
used. For now, this includes all `RelOp`s, the `YieldOp`, as well as
types tested inside `RelationType`, which covers all occurrences of
types in the test suite except for the types of `StringAttr`s, such as
in `literal "foo" : !substrait.string`.  This PR does also not change
the assembly format of attributes, which have a similar source of
redundancy that we may want to remove in follow-up work.

Signed-off-by: Ingo Müller <ingomueller@google.com>