ci(dependabot): try different syntax for module filtering (#151)
The intention of the dependabot configuration is to update LLVM weekly
and everything else daily. This was achieved with the `allow` and
`ignore` tags, which take the dependency name as an argument.
Apparently, what I tried first (`llvm-project`) does not correspond to
the expected format since we now see daily updates for LLVM. This PR
changes the syntax to the string in quotes in `.gitmodules`.

Signed-off-by: Ingo Müller <ingomueller@google.com>