build: enable `misc-use-anonymous-namespace` check of `clang-tidy` (#146)
That flag was disabled when `clang-tidy` was set up initially in #140
because it failed in too many places. This PR enables it and fixes all
instances found by `clang-tidy` such that the style will be enforced in
the future.

The PR also moves some functions from the `mlir::substrait`
namespace that were not declared in any header file into the anonymous
namespace and reorganizes the printing and parsing functions in
`Substrait.cpp`: all definitions are moved to the dedicated section
(some of which were inlined with the op member functions) and to the
anonymous namespace, all forward declarations are removed (because they
are now unnecessary), and the printer/parser functions that are op
member functions are move to the ops section of the file.

Signed-off-by: Ingo Müller <ingomueller@google.com>