style: enable and reconfigure`readability-braces-around-statements` (#143)
That flag was disabled when `clang-tidy` was set up initially in #140
because it failed in too many places. This PR enables it, such that
identifier naming will be enforced in the future. It also changes the
size of "short" statements, which don't need braces, to three lines
because all examples of two lines that the tool wanted to change seemed
perfectly readable as is to me. Finally, the PR applies the automatic
fixes such that the code is compliant with the new configuration.

Since we set the size of "short statements" to three, a few short
statements don't get braces by `clang-tidy` that acutally should
according to the LLVM Coding Standards. This commit fixes two instances
of those.

Signed-off-by: Ingo Müller <ingomueller@google.com>