ci: set up build without assertions as post-merge CI job (#160)
This PR adds another CI target that compiles the project without
assertions and that is only enabled as a post-merge check. This
addresses #158, which tracks the fact that some warnings are only raised
when a project is compiled without assertions, most notably when
variables are only used by assertions and then become unused. This is a
rare and minor issue, so the delay that would be caused by checking for
every commit of every PR would not be worth it. The post-merge check is
achieved by turning the main CI job for compiling into a `matrix` whose
value sets the flag that enables assertions in CMake and then excluding
the `ON` entry for all but `push` events (which are triggered after a PR
is "pushed" onto main during merge).

Signed-off-by: Ingo Müller <ingomueller@google.com>