feat: enable stack printing signal handler in opt builds (#162)
This PR enables the signal handler that prints a stack trace on error in
release builds (and even if assertions are disabled). Previously, the
signal handler was only installed if `NDEBUG` was set. However, that
lead to warnings about unused includes. I also don't see a good reason
why the handlers shouldn't be enabled in release mode as well. Finally,
`substrait-translate` did not have the logic for the system handlers,
yet, so this PR also adds them there.

Signed-off-by: Ingo Müller <ingomueller@google.com>