refactor: move format handling from nanobind to Python
This PR refactors the handling of the various serialization formats of
the Substrait protobuf messages, in particular, on a Python level.
Previously, the various `to_*` and `from_*` functions where implemented
in C++ using `nanobind`, which is arguable cumbersome. The new
implementation only implements a single import and export function,
repsepectively, which gets an enum argument for the format. For that
purpose, the PR also exposes that enum via `nanobind`. The previously
existing import and export functions are now implemented in Python,
where they are essentially one-liners calling the new functions.

Signed-off-by: Ingo Müller <ingomueller@google.com>