ci(deps): update Python to 3.12 and requirements to ~latest
This PR updates the Python version from 3.11 (which is three years old)
to 3.12. It also updtes all requirements managed by `pip` to the latest
available version, except for `pyarrow`. The versions of that package
above v15 do not seem to have correct type annotations or don't even
include Substrait. This also limits the Python version we upgrade to
because v15 does not have wheels for Python v3.13.

All of this only immediately affects CI, so existing versions should
continue to work, but new users following the README will also start to
use the new versions, so it's probably recommended that other developers
upgrade as well.

Signed-off-by: Ingo Müller <ingomueller@google.com>