build(nanobind): remove problematic method created by stubgen
This commits sets up a patterns file for nanobind stubgens, which is the
way of that tool to do manual modifications to the generated typing
stub, in order to remove the generated `__str__` method of an enum,
which yielded an error in the `pyright` type checks. We don't really
care about this typing information but the error blocks CI, so it's best
to just remove that function.

Signed-off-by: Ingo Müller <ingomueller@google.com>