refactor: move format handling from nanobind to Python (#167)
This PR refactors the handling of the various serialization formats of
the Substrait protobuf messages, in particular, on a Python level.
Previously, the various `to_*` and `from_*` functions where implemented
in C++ using `nanobind`, which is arguable cumbersome. The new
implementation only implements a single import and export function,
repsepectively, which gets an enum argument for the format. For that
purpose, the PR also exposes that enum via `nanobind`. The previously
existing import and export functions are now implemented in Python,
where they are essentially one-liners calling the new functions.

The PR also sets up a patterns file for nanobind stubgens, which is the
way of that tool to do manual modifications to the generated typing
stub, in order to remove the generated `__str__` method of an enum,
which yielded an error in the `pyright` type checks. We don't really
care about this typing information but the error blocks CI, so it's best
to just remove that function.

Signed-off-by: Ingo Müller <ingomueller@google.com>