ci(deps): update Python to 3.12 and requirements to ~latest (#170)
This PR updates the Python version from 3.11 (which is three years old)
to 3.12. It also updates all requirements managed by pip to the latest
available version, except for pyarrow and substrait. We don't want to
update the substrait package because it would also change the
underlying version of the Substrait spec, and we want it to use the
version we also use (v0.42.1). This restricts us to pyarrow<=15,
because higher versions package the Python proto files in a different
way. This, in turn, also limits the Python version we upgrade to
because v15 does not have wheels for Python v3.13.

All of this only immediately affects CI, so existing versions should
continue to work, but new users following the README will also start to
use the new versions, so it's probably recommended that other
developers upgrade as well.

Finally, the PR increases the timeout of the main workflow to 2.5h in
order to be long enough for scratch builds.

Signed-off-by: Ingo Müller <ingomueller@google.com>