style(python): fix and improve type annotations for export funcs (#178)
This PR fixes and improves the type annotations for the export functions
`to_(json|protobuf|text)`. The fix consists of removing the `Optional`
part for a manually defined signature that wasn't actually optional. The
improvement consists of allowing `OpView` in addition to `Operation`,
both of which can be accepted for an `MlirOperation` argument.

Signed-off-by: Ingo Müller <ingomueller@google.com>