build(bazel): migrate to bazel mode and bazel version 8.4
This commit replaces the now deprecated `WORKSPACE.bazel` with the new
approach using `MODULE.bazel` and upgrades bazel to version 8.4. The
immediate reason for this change was the fact that the current LLVM
repository does not build with the current set-up anymore. I tried a bit
to fix the old version but didn't make it work. Migrating to bazelmod
also took some time but was ultimately successful. Also, the old bazel
version was about to be end-of-life, plus the new approach with modules
*is* more elegant, so migrating is a good idea anyways.

Signed-off-by: Ingo Müller <ingomueller@google.com>