ci(pyright): suppress some type checking errors
This commits silences two kinds of type checking errors found by
pyright. Both are related to the method of extending tablegen'd ops in
Python by inheriting from the generated class with the same name. The
first one complains during import because the imported (generated) class
clashes with the class defined in the Python file, which is the whole
purpose of that class. The second one complains about the manually
defined classes overriding the `body` property of the generated base
class with a different return type but, again, the purpose of the
overriding is to provide access to the one `Block` instead of the
`Region`.

Signed-off-by: Ingo Müller <ingomueller@google.com>